+++ cpp/brain/controller/LayeredExtendedANN.cpp
@@ -23,7 +23,7 @@ LayeredExtNNController::LayeredExtNNController(std::string modelName,
   idToNeuron_ = Config->idToNeuron_;
   connections_ = Config->connections_;
 
-  unsigned int p = 0;
+  size_t p = 0;
   for (auto sensor : sensors) {
     p += sensor->inputs();
   }
@@ -50,7 +50,7 @@ LayeredExtNNController::update(const std::vector<ActuatorPtr> &actuators,
                                double step)
 {
   // Read sensor data into the input buffer
-  unsigned int p = 0;
+  size_t p = 0;
   for (auto sensor : sensors) {
     sensor->read(&inputs_[p]);
     p += sensor->inputs();
@@ -120,7 +120,7 @@ LayeredExtNNController::writeNetwork(std::ofstream &write_to)
                        v.end());
   }
   boost::adjacency_list<> graph(allNeurons_.size());
-  for (unsigned int i = 0; i < allNeurons_.size(); i++) {
+  for (size_t i = 0; i < allNeurons_.size(); i++) {
     std::vector<std::pair<std::string, NeuralConnectionPtr>> connectionsToAdd = allNeurons_[i]->getIncomingConnections();
     for (std::pair<std::string, NeuralConnectionPtr> connectionToAdd : connectionsToAdd) {
       NeuronPtr input = connectionToAdd.second
